+++ cmd/amc/main.go
@@ -1,22 +1,14 @@
 package main
 
 import (
-	"flag"
-	"fmt"
 	"os"
 
-	"github.com/Horcag/agent-machine-control/internal/buildinfo"
+	"github.com/Horcag/agent-machine-control/internal/entrypoint"
 )
 
 func main() {
-	version := flag.Bool("version", false, "print version information")
-	flag.Parse()
-
-	if *version {
-		fmt.Println(buildinfo.String("amc"))
-		return
-	}
-
-	fmt.Fprintln(os.Stderr, "amc: no command selected; use --version while the CLI is being bootstrapped")
-	os.Exit(2)
+	os.Exit(entrypoint.Run(entrypoint.Config{
+		Name:               "amc",
+		UnavailableMessage: "amc: no command selected; use --version while the CLI is being bootstrapped",
+	}, os.Args[1:], os.Stdout, os.Stderr))
 }